Add files via upload
Successfully developed a predictive model for a financial institution to identify customers at a lower risk of loan default. Additionally, I created a tailored questionnaire based on the model's insights to be used during the loan application process, enabling the institution to gather relevant information and make more informed lending decisions. The implementation of this model and questionnaire enhances the institution's ability to identify and approve clients with a reduced likelihood of default, aligning with their strategic goal of minimizing financial risks.